requirements.txt
alabaster==0.7.8
asn1crypto==0.24.0
bleach==3.1.4
cryptography==2.1.4
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-compressor==2.4
django-libsass==0.8
django-markdownx==3.0.1
django-recaptcha==2.0.6
docutils==0.14
gunicorn==20.0.4
httplib2==0.9.2
imagesize==0.7.1
Jinja2==2.10
libsass==0.19.4
Mako==1.0.7
Markdown==3.2.1
#MarkupSafe==1.0
Pillow==7.1.1
psycopg2-binary==2.8.5
pycrypto==2.6.1
python-dateutil==2.6.1
pytz==2019.3
#setuptools==46.1.3
whitenoise==5.0.1
